+++ CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Added local files policy to manifest configuration.
+- Added `local_policy` to `ManifestConfiguration`.
+- Added `file_locations` to `GetManifestResult`.
 
 ## [0.3.0] - 2026-01-14